refactor(client): move admin slice into features/admin
