+++ client/src/app/App.tsx
@@ -16,13 +16,13 @@ import Inventory from '../pages/Inventory';
 const BarcodeTools = lazy(() => import('../pages/BarcodeTools'));
 const Deliveries = lazy(() => import('../pages/Deliveries'));
 const SalesHistory = lazy(() => import('../pages/SalesHistory'));
-const UsersPage = lazy(() => import('../pages/Users'));
+const UsersPage = lazy(() => import('../features/admin/pages/Users'));
 const CustomersPage = lazy(() => import('../pages/Customers'));
 const DistributorsPage = lazy(() => import('../pages/Distributors'));
 const CategoriesPage = lazy(() => import('../pages/Categories'));
-const SettingsPage = lazy(() => import('../pages/Settings'));
+const SettingsPage = lazy(() => import('../features/admin/pages/Settings'));
 const PurchaseOrdersPage = lazy(() => import('../pages/PurchaseOrders'));
-const AuditLogPage = lazy(() => import('../pages/AuditLog'));
+const AuditLogPage = lazy(() => import('../features/admin/pages/AuditLog'));
 const PromotionsPage = lazy(() => import('../pages/Promotions'));
 const GiftCardsPage = lazy(() => import('../pages/GiftCards'));
 const StockCountPage = lazy(() => import('../pages/StockCount'));
@@ -35,8 +35,8 @@ const SegmentsPage = lazy(() => import('../pages/Segments'));
 const LayawayPage = lazy(() => import('../pages/Layaway'));
 const BundlesPage = lazy(() => import('../pages/Bundles'));
 const FeedbackPage = lazy(() => import('../pages/Feedback'));
-const BackupPage = lazy(() => import('../pages/Backup'));
-const BranchesPage = lazy(() => import('../pages/Branches'));
+const BackupPage = lazy(() => import('../features/admin/pages/Backup'));
+const BranchesPage = lazy(() => import('../features/admin/pages/Branches'));
 const StorefrontPage = lazy(() => import('../pages/Storefront'));
 const OnlineOrdersPage = lazy(() => import('../pages/OnlineOrders'));
 const ReportBuilderPage = lazy(() => import('../pages/ReportBuilder'));
