+++ client/src/features/admin/pages/AuditLog.tsx
@@ -26,13 +26,19 @@ import {
   Activity,
   X,
 } from 'lucide-react';
-import { Badge } from '../shared/ui/badge';
-import { Input } from '../shared/ui/input';
-import { Label } from '../shared/ui/label';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../shared/ui/select';
-import { useTranslation } from '../shared/i18n/index';
-import { resource } from '../shared/lib/resource';
-import { useApiQuery } from '../shared/lib/apiQuery';
+import { Badge } from '../../../shared/ui/badge';
+import { Input } from '../../../shared/ui/input';
+import { Label } from '../../../shared/ui/label';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '../../../shared/ui/select';
+import { useTranslation } from '../../../shared/i18n/index';
+import { resource } from '../../../shared/lib/resource';
+import { useApiQuery } from '../../../shared/lib/apiQuery';
 import type { LucideIcon } from 'lucide-react';
 import type { AuditEntry, User as UserRecord } from '@/types';
